# internal/docs/swagger.yaml
basePath: /api/v1
definitions:
  articles.ArticleRequest:
    properties:
      category:
        minLength: 3
        type: string
      content:
        minLength: 200
        type: string
      status:
        enum:
        - Publish
        - Draft
        - Thrash
        type: string
      title:
        minLength: 20
        type: string
    required:
    - category
    - content
    - status
    - title
    type: object
  articles.ArticleResponse:
    properties:
      category:
        type: string
      content:
        type: string
      created_at:
        type: string
      id:
        type: integer
      status:
        type: string
      title:
        type: string
      updated_at:
        type: string
    type: object
host: localhost:9876
info:
  contact: {}
  description: API untuk CRUD artikel
  title: Articles API
  version: "1.0"
paths:
  /articles/:
    get:
      description: Get paginated list of articles
      parameters:
      - description: Limit per page
        in: query
        name: limit
        type: integer
      - description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/articles.ArticleResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List all articles
      tags:
      - Articles
    post:
      consumes:
      - application/json
      description: Create a new article with title, content, category, status
      parameters:
      - description: Article Request
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/articles.ArticleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/articles.ArticleResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new article
      tags:
      - Articles
  /articles/{id}:
    delete:
      description: Delete an article by ID
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete an article
      tags:
      - Articles
    get:
      description: Get article by ID
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/articles.ArticleResponse'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get article by ID
      tags:
      - Articles
    put:
      consumes:
      - application/json
      description: 'Update article fields: title, content, category, status'
      parameters:
      - description: Article ID
        in: path
        name: id
        required: true
        type: integer
      - description: Article Request
        in: body
        name: article
        required: true
        schema:
          $ref: '#/definitions/articles.ArticleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update an article
      tags:
      - Articles
swagger: "2.0"